+++ ios/audio_waveforms/Package.swift
@@ -1,0 +1,25 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "audio_waveforms",
+    platforms: [
+        .iOS("13.0")
+    ],
+    products: [
+        .library(name: "audio-waveforms", targets: ["audio_waveforms"])
+    ],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework")
+    ],
+    targets: [
+        .target(
+            name: "audio_waveforms",
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework")
+            ]
+        )
+    ]
+)